ejer

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: astefane <astefane@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/11 17:27:42 by astefane          #+#    #+#             */
-/*   Updated: 2024/03/13 16:50:26 by astefane         ###   ########.fr       */
+/*   Updated: 2024/03/18 12:49:05 by astefane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
